// apps/web/generated/api/models/ratingResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RatingResponse {
  rating_id: string;
  customer_review_id: string;
  project_id: string;
  customer_user_id: string;
  freelancer_profile_id: string;
  score: number;
  comment: string | null;
  is_public: boolean;
}